The main and some modifications to the classes
